### 2025-07-01 #8523 @editorialbot
[PRE REVIEW]: MNE-RSA: Representational Similarity Analysis on EEG, MEG and fMRI data
**Submitting author:** @wmvanvliet (<a href="http://orcid.org/0000-0002-6537-6899">Marijn van Vliet</a>)
**Repository:** https://github.com/mne-tools/mne-rsa
**Branch with paper.md** (empty if default branch):
**Version:** 0.10
