### 2025-07-01 #8523 @wmvanvliet
@britta-wstnr would be great as an editor or reviewer for this one I think. Anyone with MNE-Python experience would be a great asset, such as [@tsbinns](https://reviewers.joss.theoj.org/reviewers/3365), [@cbrnr](https://reviewers.joss.theoj.org/reviewers/2651), and [@alexrockhill](https://reviewers.joss.theoj.org/reviewers/1359).
