### 2025-07-01 88e8a14
Could not get a hold yet of Ossi
