### 2025-07-01 2cb991f
Add full author list (#49)
* Add full author list

* Could not get a hold yet of Ossi
