### 2025-07-11 5ed6390
Update README.md (#53)
Fix links to the documentation in the README
