+++ .pre-commit-config.yaml
@@ -1,7 +1,7 @@
 repos:
   # Ruff mne-rsa
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.12.2
+    rev: v0.12.3
     hooks:
       - id: ruff
         name: ruff-lint
