### 2025-07-15 2df1dba
Add JOSS status badge (#51)
