+++ dev/_downloads/2782ebad4ad874ca6cb270c09cf9d052/03_statistics_tutorial.py
@@ -1,0 +1,560 @@
+"""
+.. _tut-statistics:
+
+Tutorial part 3: statistical analysis of group-level RSA results
+----------------------------------------------------------------
+
+In this tutorial, we'll dive into cluster-based permutation testing for group-level
+statistical analysis.
+
+The statistical test is described in `Maris & Oostenveld, 2007
+<https://doi.org/10.1016/j.jneumeth.2007.03.024>`__. This functionality is not part of
+the MNE-RSA package, but rather MNE-Python, but it is a large part of a proper RSA
+analysis so I figures any decent RSA tutorial should cover this.
+
+Cluster-based permutation testing is becoming the standard “go-to” method for testing
+for significant differences between experimental conditions in MEG and EEG studies.
+However, applying the technique correctly is not always straightforward and there are a
+lot of caveats to beware of. On top of this, the API of MNE-Python regarding these tests
+could use a bit more work. (If you are inspired by this tutorial, could you `lend a hand
+<https://github.com/mne-tools/mne-python/issues/4859>`__ with this? ❤)
+
+The purpose of cluster-based permutation testing
+------------------------------------------------
+
+When exploring differences between conditions, a good first step is to perform separate
+t-tests for each channel/sourcepoint/timepoint and visualize them. While this gives you
+a good feel of where and when significant differences may occur, one cannot always make
+solid statistical inferences based on these values. Cluster-based permutation testing is
+designed to solve two problems that arise when using mass t-tests:
+
+1. The multiple-comparisons problem: solved by using a cluster-based
+   statistic
+2. The non-Gaussian data problem: solved by using random permutations
+
+The solutions to these problems come at a hefty price though: the test can only infer
+the data is different between conditions. **The test can not be used to infer when and
+where there is a significant difference.** Researchers get this wrong a lot. The
+Fieldtrip manual even has a `dedicated page about this
+<https://www.fieldtriptoolbox.org/faq/how_not_to_interpret_results_from_a_cluster-based_permutation_test>`__
+which I encourage you to read.
+
+Now that I’ve scared you enough, let’s do this!
+
+In the cell below, update the ``data_path`` variable to point to where you have
+extracted the
+```rsa-data.zip`` <https://github.com/wmvanvliet/neuroscience_tutorials/releases/download/2/rsa-data.zip>`__
+file to.
+"""
+# ruff: noqa: E402
+# sphinx_gallery_thumbnail_number=2
+
+import mne
+
+mne.set_log_level("warning")  # Make the output less verbose
+
+# Set this to where you've extracted `rsa-data.zip` to
+data_path = "data"
+
+########################################################################################
+# Cluster permutation testing on a single channel
+# -----------------------------------------------
+#
+# In the classic EEG event-related potential (ERP) literature, statistical
+# testing between two conditions was done by first identifying one or more
+# sensors and stretches of time that seem likely to hold a significant
+# difference, then averaging the signal across those sensors and time
+# region, followed by a single t-test or ANOVA test. Time would tell
+# whether the chosen channels and time region would replicate across
+# studies.
+#
+# Let’s do this for our sensor-level RSA result that we obtained in
+# :ref:`tut-sensor-level`. I’ve ran the same analysis across multiple subjects and
+# placed it in the ``group_analysis`` subfolder of the data package you’ve downloaded.
+# Executing the call below will load them.
+
+import mne
+
+subjects = [2, 3, 4, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 17, 18, 19]
+
[…]
