### 2025-08-25 bcdd118
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.12.8 → v0.12.10](https://github.com/astral-sh/ruff-pre-commit/compare/v0.12.8...v0.12.10)
