+++ stable/_sources/functions/mne_rsa.rsa_nifti.rst.txt
@@ -1,0 +1,12 @@
+﻿mne_rsa.rsa_nifti
+=================
+
+.. currentmodule:: mne_rsa
+
+.. autofunction:: rsa_nifti
+
+.. include:: mne_rsa.rsa_nifti.examples
+
+.. raw:: html
+
+    <div style='clear:both'></div>
