+++ stable/_sources/auto_examples/source_level/plot_volume.rst.txt
@@ -1,0 +1,292 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "auto_examples/source_level/plot_volume.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_auto_examples_source_level_plot_volume.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_auto_examples_source_level_plot_volume.py:
+
+
+Source-level RSA using a searchlight on volumetric data
+-------------------------------------------------------
+
+This example demonstrates how to perform representational similarity analysis (RSA) on
+volumetric source localized MEG data, using a searchlight approach.
+
+In the searchlight approach, representational similarity is computed between the model
+and searchlight "patches". A patch is defined by a seed voxel in the source space and
+all voxels within a given radius. By default, patches are created using each voxel as a
+seed point, so you can think of it as a "searchlight" that scans through the brain.
+
+The radius of a searchlight can be defined in space, in time, or both. In this example,
+our searchlight will have a spatial radius of 1 cm. To save computation time, we will
+only perform the RSA on a single time point, but feel free to experiment with specifying
+a temporal radius and performing the RSA in time as well.
+
+The dataset will be the MNE-sample dataset: a collection of 288 epochs in which the
+participant was presented with an auditory beep or visual stimulus to either the left or
+right ear or visual field.
+
+| Authors:
+| Marijn van Vliet <marijn.vanvliet@aalto.fi>
+
+.. GENERATED FROM PYTHON SOURCE LINES 28-37
+
+.. code-block:: Python
+
+    # sphinx_gallery_thumbnail_number=2
+
+    # Import required packages
+    import mne
+    import mne_rsa
+    from nilearn.plotting import plot_stat_map
+
+    mne.set_log_level(False)  # Be less verbose
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 38-39
+
+We'll be using the data from the MNE-sample set.
+
+.. GENERATED FROM PYTHON SOURCE LINES 39-43
+
+.. code-block:: Python
+
+    sample_root = mne.datasets.sample.data_path(verbose=True)
+    sample_path = sample_root / "MEG" / "sample"
+    mri_dir = sample_root / "subjects" / "sample"
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 44-46
+
+Creating epochs from the continuous (raw) data. We downsample to 100 Hz to speed up
+the RSA computations later on.
+
+.. GENERATED FROM PYTHON SOURCE LINES 46-51
+
+.. code-block:: Python
+
+    raw = mne.io.read_raw_fif(sample_path / "sample_audvis_filt-0-40_raw.fif")
+    events = mne.read_events(sample_path / "sample_audvis_filt-0-40_raw-eve.fif")
+    event_id = {"audio/left": 1, "audio/right": 2, "visual/left": 3, "visual/right": 4}
+    epochs = mne.Epochs(raw, events, event_id, preload=True)
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 52-58
+
+It's important that the model RDM and the epochs are in the same order, so that each
[…]
