### 2025-09-29 #8523 @weiyongxu
Hi  @dstansby ! Thanks for the invitation. I am happy to review this submission. The review guidelines link in your post doesn’t seem to work. Could you share the correct link?

> 👋 [@drammock](https://github.com/drammock), [@tsbinns](https://github.com/tsbinns), [@weiyongxu](https://github.com/weiyongxu), [@alinepriscillia](https://github.com/alinepriscillia) - would you be interested in reviewing this JOSS submission? For more information on what reviewing for JOSS involves, see our [reviewer guidelines](joss.readthedocs.io/en/latest/reviewer_guidelines.html). A copy of the paper can be seen in the comments below this one. If you're interested and have any questions, feel free to reach out to me by email, or comment here.
