### 2025-09-30 #8523 @dstansby
👋 @alexrockhill, @tmedani - would you be interested in reviewing this JOSS submission? For more information on what reviewing for JOSS involves, see our [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). A [copy of the paper can be seen on GitHub at this link](https://github.com/openjournals/joss-papers/blob/joss.08523/joss.08523/10.21105.joss.08523.pdf). If you're interested and have any questions, feel free to reach out to me by email, or comment here.
