### 2025-09-30 #8523 @alexrockhill
Really ideally it would be someone outside the MNE organization. I'm still technically on the steering committee but the qualification is activity within the last year so I might not be soon. Since that's the case, I would really go with someone who doesn't have that affiliation but if you really have trouble finding someone please loop back. Maybe @apoorva6262 would fit that bill as a suggestion for replacement.
