### 2025-10-01 #8523 @wmvanvliet
Yes it is important to have reviewers outside of the MNE organization, and I'm very happy we have some!

Someone with extensive experience with the MNE ecosystem will probably have an easier time installing, using and understanding MNE-RSA and therefore miss things that that are poorly documented or would otherwise be unintuitive to someone else. On the other hand, I would personally also appreciate an additional review from such an insider, as they can point out ways in which MNE-RSA breaks convention or could integrate better with some some MNE-Python functionality that we may have overlooked.
