+++ stable/_sources/auto_examples/source_level/plot_nifti.rst.txt
@@ -1,0 +1,241 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "auto_examples/source_level/plot_nifti.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_auto_examples_source_level_plot_nifti.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_auto_examples_source_level_plot_nifti.py:
+
+
+Source-level RSA using a searchlight on fMRI data
+-------------------------------------------------
+
+This example demonstrates how to perform representational similarity analysis (RSA) on
+volumetric fMRI data, using a searchlight approach.
+
+In the searchlight approach, representational similarity is computed between the model
+and searchlight "patches". A patch is defined by a seed voxel and all voxels within a
+given radius. By default, patches are created using each voxel as a seed point, so you
+can think of it as a "searchlight" that scans through the brain. In this example, our
+searchlight will have a spatial radius of 1 cm.
+
+The dataset will be the Haxby et al. 2001 dataset: a collection of 1452 scans during
+which the participant was presented with a stimulus image belonging to any of 8
+different classes: scissors, face, cat, shoe, house, scrambledpix, bottle, chair.
+
+| Authors:
+| Marijn van Vliet <marijn.vanvliet@aalto.fi>
+
+.. GENERATED FROM PYTHON SOURCE LINES 24-34
+
+.. code-block:: Python
+
+    # sphinx_gallery_thumbnail_number=2
+
+    # Import required packages
+    import mne_rsa
+    import nibabel as nib
+    import pandas as pd
+    import tarfile
+    import urllib.request
+    from nilearn.plotting import plot_glass_brain
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 35-37
+
+We'll be using the data from the Haxby et al. 2001 set, which can be found at:
+http://data.pymvpa.org/datasets/haxby2001
+
+.. GENERATED FROM PYTHON SOURCE LINES 37-57
+
+.. code-block:: Python
+
+
+    # Download and extract data
+    fname, _ = urllib.request.urlretrieve(
+        "http://data.pymvpa.org/datasets/haxby2001/subj1-2010.01.14.tar.gz"
+    )
+    tar = tarfile.open(fname, "r:gz")
+    tar.extractall()
+    tar.close()
+
+    # Load fMRI BOLD data
+    bold = nib.load("subj1/bold.nii.gz")
+
+    # This is a mask that the authors provide. It is a GLM contrast based localizer map that
+    # extracts an ROI in the "ventral temporal" region.
+    mask = nib.load("subj1/mask4_vt.nii.gz")
+
+    # This is the metadata of the experiment. What stimulus was shown when etc.
+    meta = pd.read_csv("subj1/labels.txt", sep=" ")
+    meta["labels"] = meta["labels"].astype("category")
+
+
+
+
+
+.. rst-class:: sphx-glr-script-out
+
+ .. code-block:: none
+
[…]
