### 2025-10-07 bb95ca0
Use n_jobs in the examples
