+++ examples/sensor_level/plot_sensor_level.py
@@ -88,7 +88,7 @@ rsa_result = mne_rsa.rsa_epochs(
     temporal_radius=0.05,  # Temporal radius of the searchlight path in seconds.
     tmin=0.15,
     tmax=0.25,  # To save time, only analyze this time interval
-    n_jobs=1,  # Only use one CPU core. Increase this for more speed.
+    n_jobs=1,  # Use this to specify the number of CPU cores to use.
     n_folds=None,  # Don't use any cross-validation
     verbose=False,
 )  # Set to True to display a progress bar
