### 2025-10-08 dda4af5
add n_jobs to unit tests
