+++ tests/test_sensor_level.py
@@ -102,6 +102,9 @@ class TestEpochRDMs:
         with pytest.raises(ValueError, match="`tmax=0.1` is smaller than `tmin=0.2`"):
             next(rdm_epochs(epochs, temporal_radius=0.02, tmax=0.1, tmin=0.2))
 
+        # Paralellization across 2 CPUs
+        rdms = list(rdm_epochs(epochs, temporal_radius=0.02, n_jobs=2))
+
     def test_rdm_spatial(self):
         """Test making RDMs with a searchlight across sensors."""
         epochs = load_epochs()
@@ -128,6 +131,9 @@ class TestEpochRDMs:
         with pytest.raises(ValueError, match="`picks` are not unique"):
             next(rdm_epochs(epochs, spatial_radius=0.05, picks=["EEG 020", "EEG 020"]))
 
+        # Paralellization across 2 CPUs
+        rdms = list(rdm_epochs(epochs, spatial_radius=0.05, n_jobs=2))
+
     def test_rdm_spatio_temporal(self):
         """Test making RDMs with a searchlight across both sensors and time."""
         epochs = load_epochs()
@@ -143,6 +149,11 @@ class TestEpochRDMs:
             rdm_epochs(epochs, spatial_radius=0.05, temporal_radius=0.02, noise_cov=cov)
         )
         assert not np.allclose(rdms, rdms_whitened)
+
+        # Paralellization across 2 CPUs
+        rdms = list(
+            rdm_epochs(epochs, spatial_radius=0.05, temporal_radius=0.02, n_jobs=2)
+        )
 
 
 class TestEvokedRDMs:
