### 2025-10-08 ababbc1
add more n_jobs=2 to unit tests
