### 2025-10-08 eb00b46
Document dependencies better in the README
