### 2025-10-08 51243c7
Also state qt6 as dependency
