### 2025-10-08 7ce6c31
Expose `n_jobs` parameter to sensor-level RDM functions. (#72)
* add n_jobs parameter to rdm_epochs and rdm_evokeds

* Use n_jobs in the examples

* Fix typo

* add n_jobs to unit tests

* add more n_jobs=2 to unit tests
