### 2025-10-08 8873073
Document dependencies better in the README (#73)
* Document dependencies better in the README

* Also state qt6 as dependency
