### 2025-10-08 ec3b128
Improve statement of need by including references to other EEG/MEG analysis toolboxes (#74)
