### 2025-10-09 3d4e429
Reformulate the "Features" section of the paper
