+++ paper/paper.bib
@@ -133,3 +133,13 @@
   publisher = {bioRxiv},
   doi = {10.1101/2025.05.22.655542},
 }
+@article{Wakeman2015,
+  title = {A Multi-Subject, Multi-Modal Human Neuroimaging Dataset},
+  author = {Wakeman, Daniel G. and Henson, Richard N.},
+  year = {2015},
+  journal = {Scientific Data},
+  volume = {2},
+  number = {1},
+  pages = {150001},
+  doi = {10.1038/sdata.2015.1},
+}
