### 2025-10-10 0038b50
add comments to example
