### 2025-10-10 d071913
add citations
