+++ paper/paper.bib
@@ -143,3 +143,13 @@
   pages = {150001},
   doi = {10.1038/sdata.2015.1},
 }
+@inproceedings{Schroff2015,
+  title = {{{FaceNet}}: {{A Unified Embedding}} for {{Face Recognition}} and {{Clustering}}},
+  booktitle = {2015 {{IEEE Conference}} on {{Computer Vision}} and {{Pattern Recognition}} ({{CVPR}})},
+  author = {Schroff, Florian and Kalenichenko, Dmitry and Philbin, James},
+  year = {2015},
+  eprint = {1503.03832},
+  pages = {815--823},
+  doi = {10.1109/CVPR.2015.7298682},
+  archiveprefix = {arXiv},
+}
