### 2025-10-12 dd7acbb
Reformulate the "Features" section of the paper (#75)
* Reformulate the "Features" section of the paper

* Add example

* tweak figure a bit

* more tweaks

* Better example

* add comments to example

* add citations

* fix typo
