+++ paper/paper.bib
@@ -133,3 +133,23 @@
   publisher = {bioRxiv},
   doi = {10.1101/2025.05.22.655542},
 }
+@article{Wakeman2015,
+  title = {A Multi-Subject, Multi-Modal Human Neuroimaging Dataset},
+  author = {Wakeman, Daniel G. and Henson, Richard N.},
+  year = {2015},
+  journal = {Scientific Data},
+  volume = {2},
+  number = {1},
+  pages = {150001},
+  doi = {10.1038/sdata.2015.1},
+}
+@inproceedings{Schroff2015,
+  title = {{{FaceNet}}: {{A Unified Embedding}} for {{Face Recognition}} and {{Clustering}}},
+  booktitle = {2015 {{IEEE Conference}} on {{Computer Vision}} and {{Pattern Recognition}} ({{CVPR}})},
+  author = {Schroff, Florian and Kalenichenko, Dmitry and Philbin, James},
+  year = {2015},
+  eprint = {1503.03832},
+  pages = {815--823},
+  doi = {10.1109/CVPR.2015.7298682},
+  archiveprefix = {arXiv},
+}
