+++ stable/_sources/auto_examples/tutorials/01_sensor_level_tutorial.rst.txt
@@ -1,0 +1,1702 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "auto_examples/tutorials/01_sensor_level_tutorial.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_auto_examples_tutorials_01_sensor_level_tutorial.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_auto_examples_tutorials_01_sensor_level_tutorial.py:
+
+
+.. _tut-sensor-level:
+
+Tutorial part 1: RSA on sensor-level MEG data
+---------------------------------------------
+In this tutorial, we will perform sensor-level RSA analysis on MEG data.
+
+We will explore how representational similarity analysis (RSA) can be used to study the
+neural representational code within visual cortex. We will start with performing RSA on
+the sensor level data, followed by source level and finally we will perform group level
+statistical analysis. Along the way, we will encounter many of the functions and classes
+offered by MNE-RSA, which will always be presented in the form of links to the
+:ref:`api_documentation` which you are encouraged to explore.
+
+The dataset we will be working with today is the MEG data of the `Wakeman & Nelson
+(2015) “faces” dataset <https://www.nature.com/articles/sdata20151>`__. During this
+experiment, participants were presented with a series of images, containing:
+
+- Faces of famous people that the participants likely knew
+- Faces of people that the participants likely did not know
+- Scrambled faces: the images were cut-up and randomly put together again
+
+As a first step, you need to download and extract the dataset: `rsa-data.zip <https://github.com/wmvanvliet/neuroscience_tutorials/releases/download/2/rsa-data.zip>`__.
+You can either do this by executing the cell below, or you can do so manually. In any
+case, make sure that the ``data_path`` variable points to where you have extracted the
+`rsa-data.zip <https://github.com/wmvanvliet/neuroscience_tutorials/releases/download/2/rsa-data.zip>`__
+file to.
+
+.. GENERATED FROM PYTHON SOURCE LINES 29-46
+
+.. code-block:: Python
+
+    # ruff: noqa: E402
+    # sphinx_gallery_thumbnail_number=8
+
+    import os
+
+    import pooch
+
+    # Download and unzip the data
+    pooch.retrieve(
+        url="https://github.com/wmvanvliet/neuroscience_tutorials/releases/download/2/rsa-data.zip",
+        known_hash="md5:859c0684dd25f8b82d011840725cbef6",
+        progressbar=True,
+        processor=pooch.Unzip(members=["data"], extract_dir=os.getcwd()),
+    )
+    # Set this to where you've extracted `rsa-data.zip` to
+    data_path = "data"
+
+
+
+
+
+.. rst-class:: sphx-glr-script-out
+
+ .. code-block:: none
+
[…]
