### 2025-12-01 6680688
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.14.5 → v0.14.7](https://github.com/astral-sh/ruff-pre-commit/compare/v0.14.5...v0.14.7)
