### 2025-12-12 cce939d
Minor paper edits (#85)
