### 2025-12-12 c16a522
Cite nibabel (#86)
