+++ paper/paper.bib
@@ -153,3 +153,113 @@
   doi = {10.1109/CVPR.2015.7298682},
   archiveprefix = {arXiv},
 }
+@software{Brett2025,
+  author       = {Brett, Matthew and
+                  Markiewicz, Christopher J. and
+                  Hanke, Michael and
+                  Côté, Marc-Alexandre and
+                  Cipollini, Ben and
+                  Papadopoulos Orfanos, Dimitri and
+                  McCarthy, Paul and
+                  Jarecka, Dorota and
+                  Cheng, Christopher P. and
+                  Larson, Eric and
+                  Halchenko, Yaroslav O. and
+                  Cottaar, Michiel and
+                  Ghosh, Satrajit and
+                  Wassermann, Demian and
+                  Gerhard, Stephan and
+                  Lee, Gregory R. and
+                  Baratz, Zvi and
+                  Moloney, Brendan and
+                  Wang, Hao-Ting and
+                  Kastman, Erik and
+                  Kaczmarzyk, Jakub and
+                  Guidotti, Roberto and
+                  Daniel, Jonathan and
+                  Duek, Or and
+                  Rokem, Ariel and
+                  Mathieu Scheltienne and
+                  Madison, Cindee and
+                  Sólon, Anibal and
+                  Morency, Félix C. and
+                  Goncalves, Mathias and
+                  Markello, Ross and
+                  Riddell, Cameron and
+                  Burns, Christopher and
+                  Millman, Jarrod and
+                  Gramfort, Alexandre and
+                  Leppäkangas, Jaakko and
+                  van den Bosch, Jasper J.F. and
+                  Vincent, Robert D. and
+                  Braun, Henry and
+                  Subramaniam, Krish and
+                  Van, Andrew and
+                  Legarreta, Jon Haitz and
+                  Gorgolewski, Krzysztof J. and
+                  Raamana, Pradeep Reddy and
+                  Klug, Julian and
+                  Vos de Wael, Reinder and
+                  Nichols, B. Nolan and
+                  Baker, Eric M. and
+                  Koudoro, Serge and
+                  Hayashi, Soichi and
+                  Pinsard, Basile and
+                  Haselgrove, Christian and
+                  Hymers, Mark and
+                  Esteban, Oscar and
+                  Pérez-García, Fernando and
+                  Becq, Guillaume and
+                  Dockès, Jérôme and
+                  Oosterhof, Nikolaas N. and
+                  Amirbekian, Bago and
+                  Christian, Horea and
+                  Nimmo-Smith, Ian and
+                  Nguyen, Ly and
+                  Suter, Peter and
+                  Reddigari, Samir and
+                  St-Jean, Samuel and
+                  Panfilov, Egor and
+                  Garyfallidis, Eleftherios and
+                  Varoquaux, Gael and
+                  Newton, Joshua and
+                  Hahn, Kevin S. and
+                  Waller, Lea and
+                  Hinds, Oliver P. and
+                  Sandro and
+                  Fauber, Bennet and
+                  Dewey, Blake and
+                  Perez, Fabian and
+                  Roberts, Jacob and
+                  Poline, Jean-Baptiste and
+                  Stutters, Jon and
+                  Jordan, Kesshi and
+                  Cieslak, Matthew and
+                  Moreno, Miguel Estevan and
+                  Hrnčiar, Tomáš and
+                  Haenel, Valentin and
+                  Schwartz, Yannick and
+                  Darwin, Benjamin C and
+                  Thirion, Bertrand and
+                  Gauthier, Carl and
+                  Solovey, Igor and
+                  Gonzalez, Ivan and
+                  Palasubramaniam, Jath and
[…]
