+++ mne_rsa/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.0-dev"
+__version__ = "1.0"
 
 from .source_level import rsa_stcs, rdm_stcs, rsa_stcs_rois, rsa_nifti, rdm_nifti
 from .sensor_level import rsa_evokeds, rsa_epochs, rdm_evokeds, rdm_epochs
