### 2025-12-12 0936bc6
Remove duplicate bibliography entry
