+++ paper/paper.bib
@@ -102,16 +102,6 @@
   pages = {1--13},
   doi = {10.3389/fnins.2013.00267},
 }
-@article{Guggenmos2018,
-  title = {Multivariate Pattern Analysis for {{MEG}}: A Comparison of Dissimilarity Measures},
-  author = {Guggenmos, Matthias and Sterzer, Philipp and Cichy, Radoslaw Martin},
-  year = {2018},
-  month = jun,
-  journal = {NeuroImage},
-  volume = {173},
-  pages = {434--447},
-  doi = {10.1016/j.neuroimage.2018.02.044},
-}
 @article{Hanke2009,
   title = {{{PyMVPA}}: A Python Toolbox for Multivariate Pattern Analysis of fMRI Data},
   author = {Hanke, Michael and Halchenko, Yaroslav O. and Sederberg, Per B. and Hanson, Stephen Jos{\'e} and Haxby, James V. and Pollmann, Stefan},
