### 2025-12-12 #9148 @wmvanvliet
Is it possible to have a page break right before the code example, so all the code would fit nicely on the same page?
