### 2025-12-12 67d4306
Small fixes to the paper (#87)
* insert page break

* Spell out USA, fix useage->usage typo
