### 2025-12-12 #9148 @wmvanvliet
@Kevin-Mattheus-Moerman done! (https://github.com/mne-tools/mne-rsa/pull/87). I even managed to add the page-break.
